+++ iot/src/main/java/de/xavaro/android/iot/things/IOTDomain.java
@@ -14,7 +14,7 @@ public class IOTDomain extends IOTObject
 
     public Double fixedLatFine;
     public Double fixedLonFine;
-    public Double fixedAltFine;
+    public Float fixedAltFine;
 
     public IOTDomain()
     {
