+++ iot/src/main/java/de/xavaro/android/iot/things/IOTLocation.java
@@ -12,7 +12,7 @@ public class IOTLocation extends IOTObject
 
     public Double fixedLatFine;
     public Double fixedLonFine;
-    public Double fixedAltFine;
+    public Float fixedAltFine;
 
     public IOTLocation()
     {
